pyproject.toml
[tool.poetry]
name = "tfm"
version = "0.1.0"
description = ""
authors = ["evilmonkey19 <eperpinyapitarch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
ntc-templates = { path = "../ntc-templates", develop=true }
netmiko = "^4.3.0"
fakenos = { path = "../fakenos", develop=true}
jinja2 = "^3.1.4"
pynautobot = "^2.2.0"
wonderwords = "^2.2.0"
ttp = "^0.9.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
